Fixing bugs
Fixed player control (jumping issue), able to respawn inside game screen (because of the resolution, player was not able to see the player respawning. That's why it seems like payer was respawning outside of the screen), added End Game UI (now there will be a button which will lead to level selection menu. Last but not least, now player will respawn at the checkpoint when the lives will loss and when there will be no lives remain, the end game UI will appear.